+++ app/eventyay/webapp/video/src/locales/en.json
@@ -164,7 +164,9 @@
 	"LandingPage:sessions:featured:link": "to full schedule",
 	"LandingPage:sessions:next:header": "Sessions Happening Soon",
 	"LandingPage:sessions:next:link": "to full schedule",
-	"LandingPage:speakers:header": "Our {{speakers}} Speakers",
+	"LandingPage:rooms:header": "Active Rooms",
+	"LandingPage:rooms:live": "LIVE",
+	"LandingPage:speakers:header": "Speakers",
 	"LandingPage:speakers:link": "to full list",
 	"LandingPage:speakers:more": "and {{additional_speakers}} more",
 	"Livestream:automuted-unmute:text": "Tap to unmute",
